remove useless DistinctQueryResult
