--- sharding-core/sharding-core-execute/src/main/java/org/apache/shardingsphere/core/execute/sql/execute/result/DistinctQueryResult.java
@@ -1,167 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.core.execute.sql.execute.result;
-
-import com.google.common.base.Function;
-import com.google.common.collect.Lists;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
-import org.apache.shardingsphere.core.execute.sql.execute.row.QueryRow;
-
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.InputStream;
-import java.io.ObjectOutputStream;
-import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.Collection;
-import java.util.Iterator;
-import java.util.LinkedHashSet;
-import java.util.List;
-
-/**
- * Distinct query result.
- *
- * @author panjuan
- * @author yangyi
- * @author sunbufu
- */
-@RequiredArgsConstructor
-@Getter(AccessLevel.PROTECTED)
-public class DistinctQueryResult implements QueryResult {
-    
-    @Getter
-    private final QueryResultMetaData queryResultMetaData;
-    
-    private final Iterator<QueryRow> resultData;
-    
-    private QueryRow currentRow;
-    
-    public DistinctQueryResult(final Collection<QueryResult> queryResults, final List<String> distinctColumnLabels) throws SQLException {
-        this.queryResultMetaData = queryResults.iterator().next().getQueryResultMetaData();
-        resultData = getResultData(queryResults, distinctColumnLabels);
-    }
-    
-    private Iterator<QueryRow> getResultData(final Collection<QueryResult> queryResults, final List<String> distinctColumnLabels) throws SQLException {
-        Collection<QueryRow> result = new LinkedHashSet<>();
-        List<Integer> distinctColumnIndexes = Lists.transform(distinctColumnLabels, new Function<String, Integer>() {
-            
-            @Override
-            public Integer apply(final String input) {
-                return getColumnIndex(input);
-            }
-        });
-        for (QueryResult each : queryResults) {
-            result.addAll(getQueryRows(each, distinctColumnIndexes));
-        }
-        return result.iterator();
-    }
-    
-    private Collection<QueryRow> getQueryRows(final QueryResult queryResult, final List<Integer> distinctColumnIndexes) throws SQLException {
-        Collection<QueryRow> result = new LinkedHashSet<>();
-        while (queryResult.next()) {
-            List<Object> rowData = new ArrayList<>(queryResult.getColumnCount());
-            for (int columnIndex = 1; columnIndex <= queryResult.getColumnCount(); columnIndex++) {
-                rowData.add(queryResult.getValue(columnIndex, Object.class));
-            }
-            result.add(new QueryRow(rowData, distinctColumnIndexes));
-        }
-        return result;
-    }
-    
-    @Override
-    public final boolean next() {
-        if (resultData.hasNext()) {
-            currentRow = resultData.next();
-            return true;
-        }
-        currentRow = null;
-        return false;
-    }
-    
-    @Override
-    public Object getValue(final int columnIndex, final Class<?> type) {
-        return currentRow.getValue(columnIndex);
-    }
-    
-    @Override
-    public Object getValue(final String columnLabel, final Class<?> type) {
-        return currentRow.getValue(getColumnIndex(columnLabel));
-    }
-    
-    @Override
-    public Object getCalendarValue(final int columnIndex, final Class<?> type, final Calendar calendar) {
-        return currentRow.getValue(columnIndex);
-    }
-    
-    @Override
-    public Object getCalendarValue(final String columnLabel, final Class<?> type, final Calendar calendar) {
-        return currentRow.getValue(getColumnIndex(columnLabel));
-    }
-    
-    @Override
-    public InputStream getInputStream(final int columnIndex, final String type) {
-        return getInputStream(currentRow.getValue(columnIndex));
-    }
-    
-    @Override
-    public InputStream getInputStream(final String columnLabel, final String type) {
-        return getInputStream(currentRow.getValue(getColumnIndex(columnLabel)));
-    }
-    
-    @SneakyThrows
-    protected InputStream getInputStream(final Object value) {
-        ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
-        ObjectOutputStream objectOutputStream = new ObjectOutputStream(byteArrayOutputStream);
-        objectOutputStream.writeObject(value);
-        objectOutputStream.flush();
-        objectOutputStream.close();
-        return new ByteArrayInputStream(byteArrayOutputStream.toByteArray());
-    }
-    
-    @Override
-    public boolean wasNull() {
-        return null == currentRow;
-    }
-    
-    @Override
-    public boolean isCaseSensitive(final int columnIndex) throws SQLException {
-        return queryResultMetaData.isCaseSensitive(columnIndex);
-    }
-    
-    @Override
-    public int getColumnCount() throws SQLException {
-        return queryResultMetaData.getColumnCount();
-    }
-    
-    @Override
-    public String getColumnLabel(final int columnIndex) throws SQLException {
-        String columnLabel = queryResultMetaData.getColumnLabel(columnIndex);
-        if (null != columnLabel) {
-            return columnLabel;
-        }
-        throw new SQLException("Column index out of range", "9999");
-    }
-    
-    protected Integer getColumnIndex(final String columnLabel) {
-        return queryResultMetaData.getColumnIndex(columnLabel);
-    }
-}
